Drops Vapor imports in logging. Changes logger names.
