--- .swiftpm/xcode/xcuserdata/aleksanderlorenc.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,62 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>SwiftBox-Package.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-		<key>SwiftBoxConfig.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>1</integer>
-		</dict>
-		<key>SwiftBoxLogging.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>2</integer>
-		</dict>
-		<key>SwiftBoxMetrics.xcscheme_^#shared#^_</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>3</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>SwiftBoxConfig</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>SwiftBoxConfigTests</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>SwiftBoxLogging</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>SwiftBoxLoggingTests</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>SwiftBoxMetrics</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>SwiftBoxMetricsTests</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
